The const scale_group makes both the objects added into 'group' and the imported blender object bigger or smaller. This is useful because it makes it possible to work with lower number for other objects that are not added to the scene. Note that this doesn't prevent having to rescale css3dobjects to make them possible to see in the scene (because of the units issue with webgl and css).
Added a background layer for new CSS3D elements, but it doesn't work too well  with the other elements. Kept this in this commit just to show what happens (the BG layer of css3d is controlle by the camera but behaves independently from the main layer - which is the one with the screen).
It also does not prevennt the CSS3D objects from showing on top of WebGL, so it'll be best for now to just use one layer of CSS3D and WebGL static background.